Extract cli script from more common scripts
