+++ package.json
@@ -18,8 +18,8 @@
     "typescript": "^4.3.5"
   },
   "scripts": {
-    "start": "ts-node src/index.ts",
-    "dev": "nodemon --quiet --watch 'src/**/*.ts' --exec 'ts-node' src/index.ts",
+    "start": "ts-node src/cli/index.ts",
+    "dev": "nodemon --quiet --watch 'src/**/*.ts' --exec 'ts-node' src/cli/index.ts",
     "build": "tsc",
     "test": "jest",
     "format": "prettier --write ./src/",
