+++ src/cli/index.ts
@@ -1,8 +1,8 @@
 import { program } from "commander";
-import { parseGameStateFromFile } from "./parseConfigFile";
-import { initScreen } from "./gui";
-import { renderBoard } from "./gui";
-import { initNewGameState } from "./gameState";
+import { parseGameStateFromFile } from "../common/parseConfigFile";
+import { initScreen } from "../common/gui";
+import { renderBoard } from "../common/gui";
+import { initNewGameState } from "../common/gameState";
 import { exit } from "process";
 
 const DEFAULT_BOARD_SIZE = 19;
