+++ src/common/graph.ts
