+++ src/common/pathfinding.ts
